\name{check_rforge}
\alias{check_rforge}
\title{
  Try to detect if we are running on R-forge
}
\description{
  R-forge has a sufficiently antiquated OS version that it's
  hard to maintain binaries for it
}
\usage{
check_rforge()
}
\value{
  logical value
}
\author{
Ben Bolker
}
\examples{
check_rforge()
}
\keyword{misc}
