\name{logLik.glmm.admb}
\alias{logLik.glmm.admb}
\title{Extract Log Likelihood}
\description{
  Extract the log likelihood of a fitted \code{glmm.admb} model.
}
\usage{
\method{logLik}{glmm.admb}(object, \dots)
}
\arguments{
  \item{object}{a fitted \code{glmm.admb} model.}
  \item{\dots}{further arguments (currently ignored).}
}
\value{An object of class \code{"logLik"}.}
\author{Hans Skaug, David Fournier, and Anders Nielsen.}
\seealso{
  \code{\link{glmm.admb}}.

  \code{\link{glmmADMB-package}}.
}
\keyword{models}
