\name{plot.glmmadmb}
\alias{plot.glmmadmb}
\title{Plot Model Fit}
\description{
  Default plotting method for a fitted \code{glmmadmb} model.
  Currently, it only plots the residuals.
}
\usage{
\method{plot}{glmmadmb}(x, \dots)
}
\arguments{
  \item{x}{a fitted \code{glmmadmb} model.}
  \item{\dots}{further arguments passed to \code{plot}.}
}
\seealso{
  \code{\link{glmmadmb}}.

  \code{\link{glmmADMB-package}}.
}
\keyword{models}
