\name{summary.glmmadmb}
\alias{summary.glmmadmb}
\alias{print.summary.glmmadmb}
\title{Model Summary}
\description{Summary of a fitted \code{glmmadmb} model.}
\usage{
\method{summary}{glmmadmb}(object, \dots)
}
\arguments{
  \item{object}{a fitted \code{glmmadmb} model.}
  \item{\dots}{further arguments}
}
\seealso{
  \code{\link{glmmadmb}}
}
\keyword{models}
