\name{glmmADMB-internal}
\alias{below}
\alias{dat_write}
\alias{par_read}
\alias{pin_write}
\title{Internal glmmADMB Functions}
\description{Internal glmmADMB functions}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{internal}
