\name{logLik.glmmadmb}
\alias{logLik.glmmadmb}
\title{Extract Log Likelihood}
\description{
  Extract the log likelihood of a fitted \code{glmmadmb} model.
}
\usage{
\method{logLik}{glmmadmb}(object, \dots)
}
\arguments{
  \item{object}{a fitted \code{glmmadmb} model.}
  \item{\dots}{further arguments (currently ignored).}
}
\value{An object of class \code{"logLik"}.}
\seealso{
  \code{\link{glmmadmb}}.

  \code{\link{glmmADMB-package}}.
}
\keyword{models}
