\name{logLik.glmm.admb}
\alias{logLik.glmm.admb}
\title{logLik.glmm.admb}
\description{Compute the log likelihood value}
\usage{
\method{logLik}{glmm.admb}(object, \dots)
}
\arguments{
  \item{object}{An object (a fit) returned from \code{glmm.admb}}
  \item{\dots}{Further arguments passed on}
}
\value{An object of class 'logLik'}
\author{
  H. Skaug skaug@mi.uib.no, David Fournier otter@otter-rsch.com and
  Anders Nielsen andersn@hawaii.edu
}
\seealso{\code{\link{glmm.admb}}}
\keyword{models}
