\name{plot.glmm.admb}
\alias{plot.glmm.admb}
\title{plot.glmm.admb}
\description{
  Default plotting method for a 'glmm.admb' object. Currently it only
  plots the residuals.
}
\usage{
\method{plot}{glmm.admb}(x, \dots)
}
\arguments{
  \item{x}{An object (a fit) returned from \code{glmm.admb}}
  \item{\dots}{Further arguments passed to \code{plot}}
}
\author{
  H. Skaug skaug@mi.uib.no, David Fournier otter@otter-rsch.com and
  Anders Nielsen andersn@hawaii.edu
}
\seealso{\code{\link{glmm.admb}}}
\keyword{models}
