\name{print.glmm.admb}
\alias{print.glmm.admb}

\title{Print glmm.admb objects}
\description{

}
\usage{
print.glmm.admb(x,...)
	}

\arguments{
  \item{x}{as returned from \code{glmm.admb} }
  \item{...}{Additional arguments here an argument \code{sd_S_print=FALSE}, if set 
             \code{TRUE} the covariance matrix of random effects vector and corresponding 
             standard deviations are obtained}
}
\author{H. Skaug skaug@mi.uib.no, David Fournier otter@otter-rsch.com and 
Anders Nielsen andersn@hawaii.edu}

\seealso{\code{\link{glmm.admb}}}

\keyword{models}% at least one, from doc/KEYWORDS

