\name{print.glmm.admb}
\alias{print.glmm.admb}
\title{Print Model}
\description{Print \code{glmm.admb} object.}
\usage{
\method{print}{glmm.admb}(x,\dots)
}
\arguments{
  \item{x}{a fitted \code{glmm.admb} model.}
  \item{\dots}{further arguments. If \code{sd_S_print=TRUE}, the
    covariance matrix of random effects vector and corresponding
    standard deviations are obtained.}
}
\author{Hans Skaug, David Fournier, and Anders Nielsen.}
\seealso{
  \code{\link{glmm.admb}}.

  \code{\link{glmmADMB-package}}.
}
\keyword{models}
