\name{summary.glmm.admb}
\alias{summary.glmm.admb}

\title{summary.glmm.admb}
\description{
  Default summary method for an 'glmm.admb' object
}
\usage{
\method{summary}{glmm.admb}(object, \dots)
}

\arguments{
  \item{object}{An object (a fit) returned from \code{glmm.admb}}
  \item{\dots}{Further arguments passed on}
}
\author{H. Skaug skaug@mi.uib.no, David Fournier otter@otter-rsch.com and 
Anders Nielsen andersn@hawaii.edu}

\seealso{\code{\link{glmm.admb}}}
\keyword{models}

