\name{summary.glmm.admb}
\alias{summary.glmm.admb}
\title{Model Summary}
\description{Summary of a fitted \code{glmm.admb} model.}
\usage{
\method{summary}{glmm.admb}(object, \dots)
}
\arguments{
  \item{object}{a fitted \code{glmm.admb} model.}
  \item{\dots}{further arguments passed to \code{print.glmm.admb}.}
}
\seealso{
  \code{\link{glmm.admb}}.

  \code{\link{glmmADMB-package}}.
}
\keyword{models}
